+++ keyboard/src/main.rs
@@ -4,8 +4,11 @@ use clap::{Parser, Subcommand};
 use clap_complete::Shell;
 use log::LevelFilter;
 use std::env;
+use std::path::PathBuf;
 use std::sync::Arc;
 use syntoniq_kbd::controller::Controller;
+#[cfg(feature = "csound")]
+use syntoniq_kbd::csound;
 use syntoniq_kbd::engine;
 use syntoniq_kbd::engine::{Keyboard, SoundType};
 use syntoniq_kbd::events::Events;
@@ -35,6 +38,8 @@ enum Commands {
     Prompt(Prompt),
     /// Output the built-in keyboard configuration
     DefaultConfig,
+    /// Output the built-in Csound text file containing the instrument
+    CsoundText,
     /// Generate shell completion
     Completion {
         /// shell
@@ -68,6 +73,10 @@ struct SoundConfig {
     /// Additional option to pass to csound, e.g. --csound-arg=-odac1; repeatable
     #[arg(long)]
     csound_arg: Vec<String>,
+    /// Csound file containing the keyboard's instrument; start with `syntoniq-kbd csound-text`
+    /// and modify based on the comments.
+    #[arg(long)]
+    csound_file: Option<PathBuf>,
 }
 
 #[tokio::main]
@@ -82,6 +91,15 @@ async fn main() -> anyhow::Result<()> {
         Commands::DefaultConfig => {
             print!("{}", engine::DEFAULT_SCORE);
             return Ok(());
+        }
+        Commands::CsoundText => {
+            #[cfg(not(feature = "csound"))]
+            bail!("MIDI not requested and csound not available");
+            #[cfg(feature = "csound")]
+            {
+                print!("{}", csound::CSOUND_TEXT);
+                return Ok(());
+            }
         }
         Commands::Run { .. } | Commands::Prompt { .. } => {}
     }
@@ -107,12 +125,13 @@ async fn main() -> anyhow::Result<()> {
     } else {
         #[cfg(feature = "csound")]
         {
-            SoundType::Csound(std::mem::take(&mut sound_config.csound_arg))
+            SoundType::Csound {
+                file: sound_config.csound_file.take(),
+                args: std::mem::take(&mut sound_config.csound_arg),
+            }
         }
         #[cfg(not(feature = "csound"))]
-        {
-            bail!("MIDI not requested and csound not available");
-        }
+        bail!("MIDI not requested and csound not available");
     };
     engine::start_sound(sound_type, events_tx.clone(), events_rx.resubscribe()).await;
 
@@ -153,7 +172,9 @@ async fn main() -> anyhow::Result<()> {
             h
         }
         Commands::Prompt(_) => prompt::run(events),
-        Commands::DefaultConfig | Commands::Completion { .. } => unreachable!("already handled"),
+        Commands::DefaultConfig | Commands::CsoundText | Commands::Completion { .. } => {
+            unreachable!("already handled")
+        }
     };
 
     drop(events_tx);